avance

+++ src/main/java/com/carlos/sistema_ventas/service/CustomerService.java
@@ -1,9 +1,14 @@
 package com.carlos.sistema_ventas.service;
 
+import com.carlos.sistema_ventas.dto.customer.CustomerRequest;
+import com.carlos.sistema_ventas.dto.customer.CustomerResponse;
+
+import java.util.List;
+
 public interface CustomerService {
     List<CustomerResponse> findAll();
-    CustomerResponse findById(Integer id);
+    CustomerResponse findById(Long id);
     CustomerResponse create(CustomerRequest request);
-    CustomerResponse update(Integer id, CustomerRequest request);
-    void delete(Integer id);
+    CustomerResponse update(Long id, CustomerRequest request);
+    void delete(Long id);
 }
